refactor(core): extract ChannelScan from audio decode
Move the header-first stereo/dual-mono scan into decode/scan.rs.
Public scan_channels / ChannelScan paths are unchanged.
